# backend/admin_openapi.yaml
openapi: 3.0.3
info:
  title: 智糖小助手 - 后台管理API
  description: '智糖小助手后台管理接口文档


    包含用户管理、FAQ管理、标签管理、Prompt管理、消息记录管理等模块。'
  version: 2.0.0
  contact:
    name: 智糖团队
servers:
- url: http://127.0.0.1:8900
  description: 本地开发环境
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Token认证，格式: Bearer <token>'
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: 错误信息
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 操作成功
...
    TagDefinition:
      type: object
      properties:
        tag_key:
          type: string
        tag_name:
          type: string
        tag_type:
          type: string
        category:
          type: string
        description:
          type: string
        validation_rules:
          type: object
...
paths:
...
  /api/tag/:
    get:
      tags:
      - 标签管理
      summary: 获取用户标签
      description: 获取当前用户的标签
      operationId: getUserTags
      parameters:
      - name: category
        in: query
        schema:
          type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  tags:
                    type: object
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - 标签管理
      summary: 设置单个标签
      description: 设置或更新用户标签
      operationId: setUserTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - tag_key
              - tag_value
              properties:
                user_id:
                  type: integer
                tag_key:
                  type: string
                tag_value:
                  type: string
                source:
                  type: string
                  default: manual
      responses:
        '200':
          description: 设置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tag/definitions:
    get:
      tags:
      - 标签管理
      summary: 获取标签定义列表（分页）
      description: 获取所有可用的标签定义
      operationId: getTagDefinitions
      parameters:
      - name: category
        in: query
        schema:
          type: string
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: page_size
        in: query
        schema:
          type: integer
          default: 50
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  definitions:
                    type: array
                    items:
                      $ref: '#/components/schemas/TagDefinition'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tag/batch:
    post:
      tags:
      - 标签管理
      summary: 批量设置标签
      description: 批量设置多个标签
      operationId: batchSetTags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - tags
              properties:
                user_id:
                  type: integer
                tags:
                  type: object
                source:
                  type: string
                  default: manual
      responses:
        '200':
          description: 设置成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  total:
                    type: integer
                  success_count:
                    type: integer
                  failed_count:
                    type: integer
                  failed_tags:
                    type: array
                    items:
                      type: string
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - 标签管理
      summary: 批量删除标签
      description: 批量删除用户标签
      operationId: batchDeleteTags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                tag_keys:
                  type: array
                  items:
                    type: string
                clear_all:
                  type: boolean
                  default: false
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  deleted_count:
                    type: integer
                  total_requested:
                    type: integer
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tag/{tag_key}:
    delete:
      tags:
      - 标签管理
      summary: 删除单个标签
      description: 删除指定标签
      operationId: deleteUserTag
      parameters:
      - name: tag_key
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tag/clear:
    post:
      tags:
      - 标签管理
      summary: 清空所有标签
      description: 清空用户的所有标签
      operationId: clearAllTags
      responses:
        '200':
          description: 清空成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  deleted_count:
                    type: integer
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tag/history:
    get:
      tags:
      - 标签管理
      summary: 获取标签历史
      description: 获取标签更新历史记录
      operationId: getTagHistory
      parameters:
      - name: user_id
        in: query
        schema:
          type: integer
      - name: limit
        in: query
        schema:
          type: integer
          default: 50
      - name: tag_key
        in: query
        schema:
          type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        tag_key:
                          type: string
                        old_value:
                          type: string
                        new_value:
                          type: string
                        source:
                          type: string
                        created_at:
                          type: string
                          format: date-time
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
security:
- BearerAuth: []